first stab at making compatible with tiddlyweb 1.1
Main changes include:
    *references to bag.list_tiddlers changed to store.list_bag_tiddlers
    *all of the custom put handling code has been removed

The second change is due to a change in the way core code in
tiddlyweb.web.handler.tiddler has been modified. Previously, core code
checked specifically for a content type of text/plain or application/json.

As of tiddlyweb 1.1, this has now changed to check for the presence of an
as_tiddler function within the appropriate serializer. This means that the
custom handling code can simply be removed, with the core code now calling
the form serializer.

Other changes relate to pulling other custom code that used to be in the
post_handler into the post_tiddler_to_container function.