Don't add the 'form' extension.
Extension are only needed for those serializations which work
with GET requests. Since form only works with POST (which _requires_
a content-type header), the extension is not needed.

As a bonus this means that it is more effective to use the list
of extension in markup for links to alternate representations (which
can be GET).