Ignore title in form in serializer to avoid TiddlerFormatError
Some time ago it was discovered that it was possible to create
tiddlers with a 'title' in the 'fields' attribute. When using
the text store this would save the title onto to disk and then
cause errors when reading the data back to create a tiddler object.

Since we've already used the title field in the form to set the
title of the tiddler we can skip over it here and avoid the
error.